Move logic for time range text to `StatsTimeRangeBarViewModel`.
