+++ WooCommerce/Classes/ViewRelated/Dashboard/Stats v4/StatsTimeRangeBarView.swift
@@ -1,37 +1,4 @@
 import UIKit
-import Yosemite
-
-private extension StatsTimeRangeV4 {
-    func timeRangeText(startDate: Date, endDate: Date, timezone: TimeZone) -> String {
-        let dateFormatter = timeRangeDateFormatter(timezone: timezone)
-        switch self {
-        case .today, .thisMonth, .thisYear:
-            return dateFormatter.string(from: startDate)
-        case .thisWeek:
-            let startDateString = dateFormatter.string(from: startDate)
-            let endDateString = dateFormatter.string(from: endDate)
-            return String.localizedStringWithFormat(NSLocalizedString("%1$@-%2$@", comment: "Displays a date range for a stats interval"), startDateString, endDateString)
-        }
-    }
-
-    func timeRangeDateFormatter(timezone: TimeZone) -> DateFormatter {
-        let dateFormatter: DateFormatter
-        switch intervalGranularity {
-        case .hourly:
-            dateFormatter = DateFormatter.Charts.chartAxisDayFormatter
-        case .daily:
-            dateFormatter = DateFormatter.Charts.chartAxisDayFormatter
-        case .weekly:
-            dateFormatter = DateFormatter.Charts.chartAxisMonthFormatter
-        case .monthly:
-            dateFormatter = DateFormatter.Charts.chartAxisYearFormatter
-        default:
-            fatalError("This case is not supported: \(intervalGranularity.rawValue)")
-        }
-        dateFormatter.timeZone = timezone
-        return dateFormatter
-    }
-}
 
 /// Contains a label that displays the time range - a date, date range for a week, month, or year.
 class StatsTimeRangeBarView: UIView {
@@ -50,13 +17,8 @@ class StatsTimeRangeBarView: UIView {
     }
 
     /// Updates the label with start/end dates, time range type, and site time zone.
-    func updateDates(startDate: Date,
-                     endDate: Date,
-                     timeRange: StatsTimeRangeV4,
-                     timezone: TimeZone) {
-        label.text = timeRange.timeRangeText(startDate: startDate,
-                                             endDate: endDate,
-                                             timezone: timezone)
+    func updateUI(viewModel: StatsTimeRangeBarViewModel) {
+        label.text = viewModel.timeRangeText
     }
 }
